Improved buttons related logic in QML

+++ meditation_app/model/meditationmodel.h
@@ -10,7 +10,7 @@ class MeditationModel : public QObject
     Q_OBJECT
     Q_PROPERTY(MeditationState state READ getState NOTIFY modelStateChanged)
     Q_PROPERTY(uint16_t duration READ getDuration WRITE setDuration NOTIFY durationChanged)
-    Q_PROPERTY(uint16_t timeLeftS READ getTimeLeft WRITE setTimeLeft NOTIFY timeLeftChanged)
+    Q_PROPERTY(uint timeLeftS READ getTimeLeft WRITE setTimeLeft NOTIFY timeLeftChanged)
 
   public:
     enum class MeditationState {
@@ -27,7 +27,7 @@ class MeditationModel : public QObject
 
     MeditationState getState() const;
     uint16_t getDuration() const;
-    uint16_t getTimeLeft() const;
+    uint getTimeLeft() const;
     void setDuration(uint16_t duration);
 
   signals:
@@ -56,7 +56,7 @@ class MeditationModel : public QObject
   private:
     MeditationState state_;
     med::MeditationConfig config_;
-    uint16_t time_left_s_;
+    uint time_left_s_;
 };
 
 #endif // MEDITATIONMODEL_H
